refactor(web): migrate from styled-components to linaria
