+++ apps/web/next.config.js
@@ -2,6 +2,8 @@
 import path from 'path';
 import url from 'url';
 
+import withLinaria from 'next-with-linaria';
+
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 
 /** @type {import('next').NextConfig} */
@@ -61,5 +63,7 @@ let nextConfig = {
     ],
   },
 };
+
+nextConfig = withLinaria(nextConfig);
 
 export default nextConfig;
