+++ apps/web/package.json
@@ -9,7 +9,7 @@
     "internal:build:next": "next build",
     "internal:build:sitemap": "next-sitemap --config next-sitemap.cjs",
     "internal:compile": "tsc -p ./tsconfig.project.json",
-    "internal:compile:watch": "pnpm run internal:compile --watch --preserveWatchOutput",
+    "internal:compile:watch": "pnpm run internal:compile --watch",
     "dev": "next dev",
     "start": "next start",
     "nuke": "pnpm run nuke:artifacts && del-cli node_modules",
@@ -24,6 +24,9 @@
   },
   "dependencies": {
     "@fontsource/rubik": "^4.5.14",
+    "@linaria/babel-preset": "^4.4.5",
+    "@linaria/core": "^4.2.10",
+    "@linaria/react": "^4.3.8",
     "@pkerschbaum-homepage/commons-ecma": "workspace:*",
     "@pkerschbaum-homepage/fetch-favicon": "workspace:*",
     "@pkerschbaum-homepage/mdx": "workspace:*",
@@ -35,13 +38,13 @@
     "gray-matter": "^4.0.3",
     "next": "^13.4.6",
     "next-mdx-remote": "4.3.0",
+    "next-with-linaria": "^0.5.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-feather": "^2.0.10",
     "react-is": "^18.2.0",
     "sanitize-html": "^2.10.0",
     "sharp": "^0.32.1",
-    "styled-components": "6.0.0-rc.2-4007",
     "tiny-invariant": "^1.3.1",
     "unist-util-visit": "^4.1.2",
     "zod": "^3.21.4"
