+++ apps/web/scripts/fetch-favicons-for-articles.ts
@@ -4,7 +4,7 @@ import path from 'path';
 import { jsonUtil } from '@pkerschbaum-homepage/commons-ecma/util/json.util';
 import { fetchFaviconsForAllHrefs } from '@pkerschbaum-homepage/fetch-favicon';
 
-import { PATHS } from '#pkg/constants.js';
+import { PATHS } from '#pkg/constants-server.js';
 
 async function fetchFaviconsForAllHrefsAndWriteToFile() {
   const [postsBasenames, tidbitBasenames] = await Promise.all([
