+++ apps/web/src/app/blog/page.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import type React from 'react';
 
 import { BlogOverviewPageContent } from '#pkg/app/blog/page-content';
-import { PATHS } from '#pkg/constants.js';
+import { PATHS } from '#pkg/constants-server.js';
 import { getAllMarkdownFiles } from '#pkg/mdx/index.js';
 
 async function BlogOverviewPage() {
