+++ apps/web/src/app/page-content.tsx
@@ -1,8 +1,8 @@
 'use client';
 
+import { styled } from '@linaria/react';
 import type React from 'react';
 import { PenTool } from 'react-feather';
-import { styled } from 'styled-components';
 
 import type { MDXFile } from '@pkerschbaum-homepage/mdx/schema';
