+++ apps/web/src/components/enable-animations-after-hydration/EnableAnimationsAfterHydration.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 
-import { DataAttribute, IsAnimationEnabled } from '#pkg/constants.js';
+import { DataAttribute, IsAnimationEnabled } from '#pkg/constants-browser.js';
 import { useIsMounted } from '#pkg/utils/react.utils.jsx';
 
 export const EnableAnimationsAfterHydration: React.FC = () => {
