+++ apps/web/src/components/header/Header.tsx
@@ -1,5 +1,5 @@
+import { styled } from '@linaria/react';
 import type React from 'react';
-import { styled } from 'styled-components';
 
 export const Header: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   return (
