+++ apps/web/src/components/icon-library/IconLibrary.tsx
@@ -1,6 +1,6 @@
+import { styled } from '@linaria/react';
 import type React from 'react';
 import type { IconProps as ReactFeatherIconProps } from 'react-feather';
-import { styled } from 'styled-components';
 
 type IconProps = {
   /**
