+++ apps/web/src/components/main/Main.tsx
@@ -1,5 +1,5 @@
+import { styled } from '@linaria/react';
 import type React from 'react';
-import { styled } from 'styled-components';
 
 type MainProps = {
   children: React.ReactNode;
