+++ apps/web/src/elements/IconButton.tsx
@@ -1,4 +1,4 @@
-import { styled } from 'styled-components';
+import { styled } from '@linaria/react';
 
 import { Button, ButtonProps } from '#pkg/elements/Button.jsx';
