+++ apps/web/src/elements/Tile.tsx
@@ -1,4 +1,4 @@
-import { styled } from 'styled-components';
+import { styled } from '@linaria/react';
 
 import { Anchor } from '#pkg/elements/Anchor.jsx';
