+++ apps/web/src/favicons/favicons.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 
 import { schema_faviconsForWebsites } from '@pkerschbaum-homepage/shared-node/schema';
 
-import { PATHS } from '#pkg/constants.js';
+import { PATHS } from '#pkg/constants-server.js';
 import type { MDXParseResult } from '#pkg/mdx/index.js';
 
 type IconURLToAssociatedWebsitesMap = {
