--- apps/web/src/styles/common.styles.ts
@@ -1,18 +1,0 @@
-import { css } from 'styled-components';
-
-export const commonStyles = {
-  /**
-   * based on https://www.joshwcomeau.com/snippets/react-components/visually-hidden/
-   */
-  visuallyHidden: css`
-    position: absolute;
-    display: inline-block;
-    width: 1px;
-    height: 1px;
-    padding: 0;
-    margin: -1px;
-    overflow: hidden;
-    clip: rect(0 0 0 0);
-    border: 0;
-  `,
-};
